Add channel member/role permission functions
Adds the following functions:

- create_permission
- delete_permission

The `create_permission` function requires the `ChannelId` to edit
the permissions of and a `PermissionOverwriteType`.

The `delete_permission` function requires the `ChannelId` to edit the
permissions of `PermissionOverwriteType`.

The `PermissionOverwriteType` is a new enum to represent the id of
either a `Member` or a `Role`. Each enum variant's value is a `UserId`
or `RoleId`, respectively.

Includes a short example in the `examples` dir for how to use
`PermissionOverwriteType`.