Add support for editing user notes
This adds functionality for the endpoint to edit a user note, and for receiving
note update events. This is a note that is user-specific (a user is the only one
that can see a note they make about another user).

This is not something listed in the official API docs, but it is a feature
utilized by the desktop client.