Implement gateway v6 support
Gateway v6 is primary the message+channel overhaul and group+call support, with
some misc. things here and there.

This includes support for the new:

Group:
- channel_id
- icon
- last_message_id
- last_pin_timestamp
- name
- owner_id
- recipients (list of users)
- call

Call:
- channel_id
- message_id
- region
- ringing (list of users)
- unavailable
- voice_states

The following new events are covered:

- CALL_CREATE
- CALL_DELETE
- CALL_UPDATE
- CHANNEL_RECIPIENT_ADD
- CHANNEL_RECIPIENT_REMOVE
- CHANNEL_PINS_ACK

In addition, this includes two new types of channels:
- Group
- Private

as channels now no longer directly denote whether they are private
(`is_private`).

`Message`s now have a new value, `kind`. This is in the form of a
`MessageType` enum, with 6 varients:
- 0: Regular
- 1: GroupRecipientAddition
- 2: GroupRecipientRemoval
- 3: GroupCallCreation
- 4: GroupNameUpdate
- 5: GroupIconUpdate

This differentiation is necessary as, for messages like "X removed Y
from the group.", there is no actual message content. This content is
generated server-side for types 1 through 5.

An overhaul of voice support is necessary due to the voice module
relying on having a `ServerId`, when groups and group calls have only a
`ChannelId`. This introduces a new `VoiceType` enum, for denoting which
is which. This enum has a few helpful impls on it to help with reducing
breakage in userland. In the voice module, passing a `ServerId` (server
voice channel) _or_ a `ChannelId` is possible via some `From` impls, so
that creating a `VoiceType` is not necessary.

`UserServerSettings` now do not _necessarily_ have a ServerId, as groups
have their own settings user-side.

Due to the aforementioned `VoiceType` impls, voice breakage is very
minimal, as can be seen in the `examples` dir diffs.