Parse integer Game types within strings
The gateway does not validate the value of a Game. Any field can be of
any type, more or less. Although we might not care about a Game type
being a float, checking for a string would fix some decoding errors, as
other libraries might erroneously send the Game's type field as an
integer within a string.

This will fix errors like:

```
[Warning] Receive error: Decode("Expected valid GameType", "1")
[Warning] Receive error: Decode("Expected valid GameType", "0")
```