port remaining libs
Bitflags is now a struct not an enum. A far as I can tell there's no way to avoid the Permissions::
Port to new websocker opcode API. Should this be using a match on the message type instead?
Port to new request multipart API. The new file part says it guesses at the mime type and filename. Should we look for a way to assure these are correct as they used to be?